+++ packages/react/src/VideoPlayer/components/FullScreenButton/index.ts
