+++ packages/react/src/VideoPlayer/components/IconControl/index.ts
