+++ packages/react/src/VideoPlayer/components/PlayPauseButton/index.ts
