--- packages/react/src/VideoPlayer/components/controls/VolumeControl/VolumeControl.tsx
@@ -1,22 +1,0 @@
-import React from 'react'
-import {Range} from '../../range'
-import {type ControlsProps} from './../Controls'
-import styles from '../../../VideoPlayer.module.css'
-
-type VolumeControlProps = Pick<ControlsProps, 'volume' | 'setVolume'>
-
-export const VolumeControl = ({volume, setVolume}: VolumeControlProps) => (
-  <Range
-    type="range"
-    min="0"
-    max={1}
-    step={0.001}
-    onInput={e => {
-      setVolume(e.currentTarget.valueAsNumber)
-    }}
-    className={styles.VideoPlayer__volumeBar}
-    value={volume}
-    a11yStep={0.1}
-    name="Volume"
-  />
-)
