+++ packages/react/src/VideoPlayer/components/index.ts
@@ -1,4 +1,11 @@
-export * from './captions'
-export * from './controls'
-export * from './range'
-export * from './tooltip'
+export * from './Captions'
+export * from './CCButton'
+export * from './Controls'
+export * from './FullScreenButton'
+export * from './IconControl'
+export * from './MuteButton'
+export * from './PlayPauseButton'
+export * from './Range'
+export * from './SeekControl'
+export * from './Tooltip'
+export * from './VolumeControl'
