+++ packages/react/src/VideoPlayer/components/range/Range.tsx
@@ -1,11 +1,11 @@
 import React, {useState, useEffect} from 'react'
 import clsx from 'clsx'
 import {useId} from '@reach/auto-id'
-import {Tooltip} from '../index'
 
+import {Tooltip} from '../'
 import styles from '../../VideoPlayer.module.css'
 
-type RangeProps = {
+export type RangeProps = {
   tooltip?: boolean
   tooltipFormatter?: (value: number | string | readonly string[]) => string
   max?: number
